test(store): share "freshPlatform" to make it reusable
