+++ packages/store/internals/testing/index.ts
@@ -1,2 +1,3 @@
 export { NgxsTestBed } from './src/ngxs.setup';
 export { NgxsTesting } from './src/symbol';
+export { freshPlatform } from './src/fresh-platform';
